More improvements
Simplified some data types in the machine model
Finished requests from the Linux side
Tested and fixed more requests
Added temporary wrapper for status responses to feed data into the machine model

+++ DuetAPI/Commands/Code/Command.cs
@@ -71,7 +71,7 @@ namespace DuetAPI.Commands
         /// <summary>
         /// File position in bytes (optional)
         /// </summary>
-        public uint? FilePosition { get; set; }
+        public long? FilePosition { get; set; }
         
         /// <summary>
         /// Retrieve the parameter whose letter equals c. Note that this look-up is case-sensitive!
